requirements.txt
uvicorn
fastapi[standard]
typing
pydantic
# pymongo
# dnspython
# mysqlclient
mysql-connector-python
